Crear pantallas

+++ src/main/java/com/example/transporte/models/dao/IEnvioDao.java
@@ -3,5 +3,5 @@ package com.example.transporte.models.dao;
 import com.example.transporte.models.entity.Envio;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IEnvioDao extends CrudRepository<Envio, Long> {
+public interface IEnvioDao extends CrudRepository<Envio, Integer> {
 }
