+++ src/main/java/com/example/transporte/models/dao/ITipoProductoDao.java
@@ -3,5 +3,5 @@ package com.example.transporte.models.dao;
 import com.example.transporte.models.entity.TipoProducto;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ITipoProductoDao extends CrudRepository<TipoProducto, Long> {
+public interface ITipoProductoDao extends CrudRepository<TipoProducto, Integer> {
 }
