+++ src/main/java/com/example/transporte/models/dao/IUsuarioDao.java
@@ -3,7 +3,7 @@ package com.example.transporte.models.dao;
 import com.example.transporte.models.entity.Usuario;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IUsuarioDao extends CrudRepository<Usuario, Long> {
+public interface IUsuarioDao extends CrudRepository<Usuario, Integer> {
 
     public Usuario findByNombre(String nombre);
 }
